board: adi: Add common support for the EZKIT and EZLITE som carrier boards
This adds common support for the SOM-CRR-EZKIT and SOM-CRR-EZLITE boards
from Analog Devices. These boards can socket multiple SOMs carrying
different processors and provide functionality that is not included on the
SOMs such as ethernet phys and "soft config" switches, which are GPIOs that
control which peripherals are active or connected to board headers.

This support is required for future commits which will introduce support
for the various SOMs that plug into these boards.

Co-developed-by: Ian Roberts <ian.roberts@timesys.com>
Signed-off-by: Ian Roberts <ian.roberts@timesys.com>
Co-developed-by: Nathan Barrett-Morrison <nathan.morrison@timesys.com>
Signed-off-by: Nathan Barrett-Morrison <nathan.morrison@timesys.com>
Signed-off-by: Greg Malysa <greg.malysa@timesys.com>