[to submit later] arm: mach-sc5xx: Remove manual bss_clear
The arm library includes an implementation of bss_clear that is already
called from crt0.S. This re-clearing of BSS should not be performed in
the machine code and must be removed.

internal note: test this separately to 100% double check, when submitted
this will be merged with the previous commit as a general cleanup of
inappropriate mach/soc code that was included in the earlier submission.