docker: add ADI build-related tools to docker image
ADI processors require packaging with the ldr utility for bootrom
compatibility. Normally this is available as part of our yocto-derived
toolchain but it is not a part of any other premade toolchain, so it is
otherwise unavailable in the docker image for CI. This patch adds a
source build from the ADI maintained github repository. The same ldr
tool is used for arm and arm64 for all of our boards with names adjusted
to match the expected $(CROSS_COMPILE) for these boards.

Signed-off-by: Greg Malysa <malysagreg@gmail.com>

---

Series-to: uboot
Series-prefix: RFC

Commit-notes:
Hi Tom, I'm not sure exactly how changes to the docker image for uboot
testing work so I wanted to get feedback on my approach here before
including it in a larger series that will include the ADI boards. I
think that maybe 10 years ago the ldr tool was available as part of some
toolchain tarballs at least, but it is gone from everywhere except the
ADI forks now. So in order to build the sc5xx boards in CI, we need to
add the ldr tool. I've built an image based on the resulting docker file
and -next and then hooked it up to my gitlab runner to test things out
and this is compatible with our upcoming patches as well for a full
clean CI run. Please let me know if I've made some mistakes or if this
isn't the right way to submit changes for the docker image.
END