mach-sc5xx: Add binman support
Binman is optionally supported for Analog Devices sc5xx if yocto is not
being used to create and assemble system images. The spl LDR is
generated locally but other artifacts such as kernel FIT image, stage 2
bootloader LDR, and root file system are built externally and must be
supplied to binman if used.

Binman is enabled by selecting the SC5XX_USE_BINMAN config symbol and
the image structure is initially defined in
arch/arm/dts/sc5xx-binman.dtb so that it is not included in the runtime
device tree.

Signed-off-by: Greg Malysa <malysagreg@gmail.com>
Signed-off-by: Utsav Agarwal <utsav.agarwal@analog.com>