mach-sc5xx: Introduce Kconfig symbols for image addresses
Add Kconfig symbols to parameterize the SPI flash layout used in a
default-ish configuration. This adds more flexibility to the default ADI
environment, enabling customers with boards based on but not identical
to an ezkit to reuse more of the infrastructure. Furthermore it allows
for yocto (the expected default) or binman (to be introduced in this
series) to configure or use the flash layout based on a single
definition of all of the parameters when creating an image.

Signed-off-by: Greg Malysa <malysagreg@gmail.com>
Signed-off-by: Utsav Agarwal <utsav.agarwal@analog.com>

---

Series-to: uboot
Series-cc: adsp-linux@analog.com

Cover-letter:
sc5xx: Add complete board support for all ADI SC5xx boards
This series adds the final pieces to enable mainline U-Boot to build and
boot all Analog Devices SC5xx SoCs and supports the associated carrier
board options. At this point it should be viable for new users for these
platforms to start with the latest version of U-Boot rather than our
vendor fork, however some features (such as OSPI support and falcon
boot) remain unavailable until we are able to unify our implementations
with the mainline implementations.
END

Cover-changes: 2
- Rebased series to latest u-boot master
- Replaced SC_BOOT_MODE Kconfig with SC_BCODE
- Use Kconfig to restrict choices between EZKIT and EZLITE carriers when
  appropriate
- Separate dtsi for revD and revE of sc598 SOM. RevE is the default but
  revD is included so that any customers with older hardware can boot it
  with a single change to the .dts
- Fixed GPIO naming and polarity to use only positive names (such as
  gige-reset rather than ~gige-reset) and set the active low/high flag
  appropriately, with all code references updated to reflect logical
  values rather than voltage levels

Cover-changes: 3
- CI results: https://gitlab.com/gmalysa/lnxdsp-u-boot/-/pipelines/2205787455